requirements.txt
# COVID-19 Classification Project Dependencies
# Author: Jordanaftermidnight

# Core Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Data Processing & Analysis
numpy>=1.21.0
pandas>=1.3.0
pillow>=11.3.0
opencv-python>=4.5.0
urllib3>=2.5.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine Learning
scikit-learn>=1.0.0

# Web Interfaces
flask>=2.0.0
streamlit>=1.40.0

# Explainable AI
grad-cam>=1.4.0

# Development & Analysis
jupyter>=1.0.0

# Security Updates
pip>=25.0